pipecat-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pipecat-ai[livekit,deepgram,cartesia,openai,grok]==0.0.104  # Includes audioop-lts for py3.13
deepgram-sdk==4.7.0
cartesia==2.0.17
openai==2.17.0
websockets==13.1
python-dotenv==1.0.1
supabase==2.9.1
httpx==0.27.2
num2words==0.5.14
sentry-sdk[fastapi,loguru]==2.25.1
pytest==8.3.3
pytest-asyncio==0.24.0
# Speaker verification (Resemblyzer). torch is installed via the CPU wheel
# index in the Dockerfile to avoid pulling the >1 GB CUDA build.
numpy==1.26.4
torch==2.5.1
resemblyzer==0.1.4
